rename file back to `timer.py`
